# REGR: Performance of DataFrame.stack where columns are not a MultiIndex